% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detectMultilayerCommunities.R
\name{detectMultilayerCommunities}
\alias{detectMultilayerCommunities}
\title{Multilayer community detection using Louvain Algorithm}
\usage{
detectMultilayerCommunities(ml, omega = NULL, seed = 123)
}
\arguments{
\item{ml}{A `buildMultilayerNetworkResult` object
from `buildMultilayerNetwork()`.}

\item{omega}{Inter-layer coupling strength (if overriding ml$params$omega).}

\item{seed}{Random seed for reproducibility.}
}
\value{
An object of class `detectMultilayerCommunitiesResult`:
\itemize{
  \item \code{stateMembership} : data.frame of gene × layer to community
  \item \code{geneMembership}: consensus across layers
  \item \code{Q}: modularity score
}
}
\description{
This function runs Louvain modularity optimization on a multilayer network.
It mirrors the approach of Russell et al. (2023), which used an iterated
GenLouvain function from a matlab github toolkit. tissue-specific
co-expression layers.
}
\examples{
\dontrun{
# Using dummy datasets to keep the example code in the functions
# fast and consistent
genes <- paste0("Gene", 1:10)

Brain  <- matrix(rnorm(10 * 5), nrow = 10,
                 dimnames = list(genes, paste0("B",1:5)))
Heart  <- matrix(rnorm(10 * 5), nrow = 10,
                 dimnames = list(genes, paste0("H",1:5)))
Liver  <- matrix(rnorm(10 * 5), nrow = 10,
                 dimnames = list(genes, paste0("L",1:5)))

adjList <- list(Brain = Brain, Heart = Heart, Liver = Liver)

ml <- buildMultilayerNetwork(adjList, threshold = 0.05, omega = 0.5)
comm <- detectMultilayerCommunities(ml)
}

}
