% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{GTExHeartTrimmed}
\alias{GTExHeartTrimmed}
\title{Heart tissue TPM gene expression data from GTEx trimmed to a reduced size}
\format{
A SummaryExperiment object with:
\describe{
 \item{TPM assay}{A matrix of size 5000 x 60 for genes x samples}
 \item{colData}{The sample names}
 \item{rowData}{The gene names}
}
}
\source{
GTEx Portal from the Broad Institute of MIT and Harvard.
}
\usage{
GTExHeartTrimmed
}
\description{
Filtered expression data from GTEx, a comprehensive RNA-seq dataset of
adult human tissue. A random 60 samples were chosen to reduce size.
Genes were filtered to the ones that code proteins, and only the top
5000 variable genes were maintained to reduce size.
}
\examples{
\dontrun{
 GTExHeartTrimmed
}
}
\keyword{datasets}
