% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotCoexpressionNetworks.R
\name{plotCoexpressionNetworks}
\alias{plotCoexpressionNetworks}
\title{Plot a multilayered graph in a neighborhood around a seed gene}
\usage{
plotCoexpressionNetworks(
  ml,
  seedGene,
  maxGenes = 10,
  k = 1,
  legendPosition = c("topleft", "topright", "bottomleft", "bottomright"),
  seed = 123
)
}
\arguments{
\item{ml}{A `buildMultilayerNetworkResult` object
from `buildMultilayerNetwork()`.}

\item{seedGene}{A character vector describing the gene ID to
center the neighborhood on. Ensure this gene is in the network.}

\item{maxGenes}{An integer representing the maximum number of genes to
include in the plot, excluding the seed gene. Default = 10.}

\item{k}{An integer representing maximum distance by which a gene is defined
as the `seedGene`'s neighbour. Default = 1 (direct neighbors).}

\item{legendPosition}{A character vector describing the position of the
legend in the figure. This is for custom positioning when the graph extends
into the legend. One of "topleft", "topright", "bottomleft", "bottomright".
Defailt = "topleft"}

\item{seed}{Random seed for reproducibility. Default = 123.}
}
\value{
Invisibly returns the igraph object used for plotting.
}
\description{
This function will filter for `maxGenes` genes surrounding the `seedGene`in
the multilayered network. Then it will plot these genes into a graph,
with genes as the nodes and the coexpression relationships as the edges,
with relationships from different layers represented in different colors
and the width of the edges represent the strength of the coexpression.
The filtering occurs in two steps, first neighbours are aggregated based
on being less than `k` distance from the `seedGene`, then this list is
filtered for the top `maxGenes` by gene connectivity to retain maximum
information
}
\examples{
\dontrun{
# Using example data from GTEx, trimmed to be lightweight and usable for
# examples
?GTExBrainTrimmed
?GTExHeartTrimmed
?GTExLiverTrimmed
# Get the coexpression data for each dataset
brainNet <- developCoexpressionNetwork(
               SummarizedExperiment::assay(GTExBrainTrimmed, "tpm"),
               corMethod = "pearson")
heartNet <- developCoexpressionNetwork(
               SummarizedExperiment::assay(GTExHeartTrimmed, "tpm"),
               corMethod = "pearson")
liverNet <- developCoexpressionNetwork(
               SummarizedExperiment::assay(GTExLiverTrimmed, "tpm"),
               corMethod = "pearson")

# Combine them together to prepare them for the multilayer network
layers <- list(Brain = brainNet$adjacency_matrix,
               Heart = heartNet$adjacency_matrix,
               Liver = liverNet$adjacency_matrix)
# Find the common genes
commonGenes <- Reduce(intersect, list(
                      rownames(brainNet$adjacency_matrix),
                      rownames(heartNet$adjacency_matrix),
                      rownames(liverNet$adjacency_matrix)
                     ))
#Building the multilayered network
multilayeredNetwork <- buildMultilayerNetwork(layers,
                                              genes = commonGenes,
                                              matchGenes = TRUE,
                                              omega = 0.5,
                                              threshold = 0.05)

MultiCoex::plotCoexpressionNetworks(multilayeredNetwork,
                                    seedGene = "ENSG00000210082.2",
                                    maxGenes = 10,
                                    k = 1,
                                    legendPosition = "topright")

}

}
\references{
Silva, A. (2022; revised 2025) TestingPackage: An Example R Package For BCB410H.
Unpublished. URL https://github.com/anjalisilva/TestingPackage.

Csárdi, G., & Nepusz, T. (2006). The igraph software package for complex
network research. InterJournal, Complex Systems, 1695. https://igraph.org

Antonov, M., Csárdi, G., Horvát, S., Müller, K., Nepusz, T., Noom, D.,
Salmon, M., Traag, V., Welles, B. F., & Zanini, F. (2023). igraph enables
fast and robust network analysis across programming languages.
arXiv:2311.10260. https://doi.org/10.48550/arXiv.2311.10260

Csárdi, G., Nepusz, T., Traag, V., Horvát, S., Zanini, F., Noom, D.,
Müller, K., Schoch, D., & Salmon, M. (2025). igraph: Network analysis and
visualization in R (Version 2.2.1). https://doi.org/10.5281/zenodo.7682609
https://CRAN.R-project.org/package=igraph
}
