% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runMultiCoex.R
\name{runMultiCoex}
\alias{runMultiCoex}
\title{Launch Shiny App for MultiCoex}
\usage{
runMultiCoex()
}
\value{
No return value but will open up a Shiny page.
}
\description{
A function that launches the Shiny app for MultiCoex, an interactive
application to run the package. The code has been placed in
\code{./inst/shiny-scripts}.
}
\examples{
\dontrun{

MultiCoex::runMultiCoex()
}

}
\references{
Grolemund, G. (2015). Learn Shiny - Video Tutorials.
\href{https://shiny.rstudio.com/tutorial/}{Link}

Silva, A. (2022; revised 2025) TestingPackage: An Example R Package For BCB410H.
Unpublished. URL https://github.com/anjalisilva/TestingPackage.

Chang, W., Cheng, J., Allaire, J. J., Sievert, C., Schloerke, B., Xie, Y.,
Allen, J., McPherson, J., Dipert, A., & Borges, B. (2024). Shiny:
Web application framework for R (Version 1.8.0). https://shiny.rstudio.com/
}
