% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/developCoexpressionNetwork.R
\name{developCoexpressionNetwork}
\alias{developCoexpressionNetwork}
\title{Develop a co-expression network from gene expression data}
\usage{
developCoexpressionNetwork(
  dataset,
  assayName = "tpm",
  TPMNormalize = FALSE,
  geneLengths = NULL,
  logScale = FALSE,
  corMethod = c("spearman", "pearson", "biweight"),
  minExp = 10,
  varianceFilter = FALSE,
  netType = "signed",
  seed = 123
)
}
\arguments{
\item{dataset}{A numeric matrix or data frame of gene expression values,
with genes in rows and samples in columns. Must have rownames (gene IDs)
and ideally column names (sample IDs).}

\item{assayName}{A string to refer to the name of the assay for the dataset
if the dataset is of type SummarizedExperiment. If no processing is done,
this name must be "tpm"}

\item{TPMNormalize}{A logical. If TRUE, performs TPM normalization assuming
input data are raw read counts. Default = FALSE.}

\item{geneLengths}{A vector of gene lengths in the order of the genes in the
dataset.Must be provided if TPM normalize is set to TRUE. Default = NULL}

\item{logScale}{Logical. If TRUE, log2-transforms the data
(after adding a pseudocount of 1). Recommended for TPM/FPKM data.
Default = FALSE.}

\item{corMethod}{Correlation method used to compute co-expression
similarity. One of "spearman", "pearson", or "biweight".
Default = "spearman".}

\item{minExp}{Minimum expression threshold (passed to
BioNERO::exp_preprocess()).
Genes with mean expression below this threshold are filtered out.}

\item{varianceFilter}{A logical representing whether to filter low-variance
genes during preprocessing. Default = FALSE.}

\item{netType}{Network type for BioNERO::exp2gcn(). Options: "signed",
"unsigned", or "signed hybrid". Default = "signed".}

\item{seed}{Random seed for reproducibility. Default = 123.}
}
\value{
A list (BioNERO network object) containing:
\itemize{
  \item \code{adjacency_matrix}: adjacency matrix of gene
  co-expression values.
  \item \code{MEs}: Data frame of module eigengenes, with samples in rows,
  and module eigengenes in columns.
  \item \code{genes_and_modules}: Data frame indicating the genes and the
  modules to which they belong.
  \item \code{kIN}: Data frame of degree centrality for each gene.
  \item \code{correlation_matrix}: Numeric matrix with pairwise correlation
  coefficients between genes
  \item \code{correlation_matrix}:Numeric matrix with pairwise correlation
  coefficients between genes.
  \item \code{params}: List with network inference parameters
  passed as input.
  \item \code{dendro_plot_objects}: List with objects to plot the dendrogram
}
}
\description{
This function streamlines the process of converting gene expression data
into a processed co-expression network using the BioNERO framework.
It automatically checks and formats the input dataset,
performs optional TPM normalization and log-transformation, determines
an optimal soft-thresholding power to achieve scale-free topology,
and constructs a gene co-expression network.
}
\examples{
\dontrun{
# Using example data from GTEx, trimmed to be lightweight and usable for
# examples
?GTExBrainTrimmed
net <- developCoexpressionNetwork(
   dataset = GTExBrainTrimmed,
   TPMNormalize = FALSE,
   logScale = FALSE,
   corMethod = "pearson",
   minExp = 10,
   varianceFilter = FALSE,
)

head(net$adjacency_matrix)

}


}
\references{
Langfelder, P., & Horvath, S. (2008). WGCNA: An R package for weighted
correlation network analysis. BMC Bioinformatics, 9, 559.
https://doi.org/10.1186/1471-2105-9-559

Almeida-Silva, F., & Venancio, T. M. (2022). BioNERO: An all-in-one
R/Bioconductor package for comprehensive and easy biological network
reconstruction. Functional & Integrative Genomics, 22, 131–136.
https://doi.org/10.1007/s10142-021-00821-9

The GTEx Consortium. (2020). The GTEx Consortium atlas of genetic regulatory
effects across human tissues. Science, 369(6509), 1318–1330.
https://doi.org/10.1126/science.aaz1776

Silva, A. (2022; revised 2025) TestingPackage: An Example R Package For BCB410H.
Unpublished. URL https://github.com/anjalisilva/TestingPackage.

Morgan, M., Obenchain, V., Hester, J., & Pagès, H. (2025).
SummarizedExperiment: A container for matrix-like assays (Version 1.38.1).
https://doi.org/10.18129/B9.bioc.SummarizedExperiment
https://bioconductor.org/packages/SummarizedExperiment
}
