% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotCoexpressionNetworks.R
\name{plotCoexpressionNetworks}
\alias{plotCoexpressionNetworks}
\title{Plot a multilayered graph in a neighborhood around a seed gene}
\usage{
plotCoexpressionNetworks(
  ml,
  seedGene,
  maxGenes = 10,
  k = 1,
  legendPosition = c("topleft", "topright", "bottomleft", "bottomright"),
  seed = 123
)
}
\arguments{
\item{ml}{A `buildMultilayerNetworkResult` object
from `buildMultilayerNetwork()`.}

\item{seedGene}{A character vector describing the gene ID to
center the neighborhood on. Ensure this gene is in the network.}

\item{maxGenes}{An integer representing the maximum number of genes to
include in the plot, excluding the seed gene. Default = 10.}

\item{k}{An integer representing maximum distance by which a gene is defined
as the `seedGene`'s neighbour. Default = 1 (direct neighbors).}

\item{legendPosition}{A character vector describing the position of the
legend in the figure. This is for custom positioning when the graph extends
into the legend. One of "topleft", "topright", "bottomleft", "bottomright".
Defailt = "topleft"}

\item{seed}{Random seed for reproducibility. Default = 123.}
}
\value{
Invisibly returns the igraph object used for plotting.
}
\description{
This function will filter for `maxGenes` genes surrounding the `seedGene`in
the multilayered network. Then it will plot these genes into a graph,
with genes as the nodes and the coexpression relationships as the edges,
with relationships from different layers represented in different colors
and the width of the edges represent the strength of the coexpression.
The filtering occurs in two steps, first neighbours are aggregated based
on being less than `k` distance from the `seedGene`, then this list is
filtered for the top `maxGenes` by gene connectivity to retain maximum
information
}
\examples{
\dontrun{
# Using dummy datasets to keep the example code in the functions
# fast and consistent
genes <- paste0("Gene", 1:10)

Brain  <- matrix(rnorm(10 * 5), nrow = 10,
                 dimnames = list(genes, paste0("B",1:5)))
Heart  <- matrix(rnorm(10 * 5), nrow = 10,
                 dimnames = list(genes, paste0("H",1:5)))
Liver  <- matrix(rnorm(10 * 5), nrow = 10,
                 dimnames = list(genes, paste0("L",1:5)))

adjList <- list(Brain = Brain, Heart = Heart, Liver = Liver)

ml <- buildMultilayerNetwork(adjList, threshold = 0.05, omega = 0.5)
comm <- detectMultilayerCommunities(ml)
plotCoexpressionNetworks(ml, "Gene4",
   maxGenes = 5)
}



}
