% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotCoexpressionNetworks.R
\name{PlotCoexpressionNetworks}
\alias{PlotCoexpressionNetworks}
\title{Plot a single community from one layer (simple)}
\usage{
PlotCoexpressionNetworks(ml, comm, community_id, layer, seed = 123)
}
\arguments{
\item{ml}{mcx_multilayer from build_multilayer(); uses ml$layer_adj}

\item{comm}{community result (e.g., from detect_genlouvain); uses $state_membership}

\item{community_id}{integer community label to plot}

\item{layer}{character; which layer (must be in ml$layer_names)}

\item{seed}{Random seed for reproducibility.}
}
\value{
invisibly returns the igraph object used for plotting
}
\description{
Plot a single community from one layer (simple)
}
